refactor(test/sdk): add run-e2e-tune-api.py.
