+++ test/e2e/v1beta1/scripts/gh-actions/build-load.sh
@@ -34,6 +34,7 @@ TAG="e2e-test"
 VERSION="v1beta1"
 CMD_PREFIX="cmd"
 SPECIFIED_DEVICE_TYPE_IMAGES=("enas-cnn-cifar10-cpu" "darts-cnn-cifar10-cpu" "pytorch-mnist-cpu")
+DEFAULT_IMAGE_FOR_TUNE="docker.io/tensorflow/tensorflow:2.13.0"
 
 IFS="," read -r -a TRIAL_IMAGE_ARRAY <<< "$TRIAL_IMAGES"
 IFS="," read -r -a EXPERIMENT_ARRAY <<< "$EXPERIMENTS"
@@ -161,6 +162,10 @@ echo -e "\nBuilding training container images..."
 for name in "${TRIAL_IMAGE_ARRAY[@]}"; do
   run "$name" "examples/$VERSION/trial-images/$name/Dockerfile"
 done
+
+# Testing image for tune function
+echo -e "\nPulling testing image for tune function..."
+docker pull $DEFAULT_IMAGE_FOR_TUNE
 
 echo -e "\nCleanup Build Cache...\n"
 docker buildx prune -f
